Make warp.panorama() much faster.
ch3_panorama now runs in less than 1 minute, instead of > 8min (after
which I killed it). The output should be the same as before. Do this by
doing the perspective image mapping using PIL instead of scipy.

Also fix a FIXME about x and y being swapped (since this wasn't
mentioned in the bug, this was pretty confusing to me), since that's no
longer necessary.